+++ providers/openshift_create_master.rb
@@ -5,6 +5,7 @@
 # Copyright (c) 2015 The Authors, All Rights Reserved.
 
 require 'openssl'
+require 'resolv'
 
 use_inline_resources
 provides :openshift_create_master if defined? provides
@@ -44,7 +45,7 @@ action :create do
     template new_resource.master_file do
       source 'master.yaml.erb'
       variables(
-        erb_corsAllowedOrigins: new_resource.origins,
+        erb_corsAllowedOrigins: new_resource.origins + [Resolv.getaddress(node['cookbook-openshift3']['openshift_common_public_hostname'])] + node['cookbook-openshift3']['openshift_common_svc_names'],
         standalone_registry: new_resource.standalone_registry,
         erb_master_named_certificates: named_certificates,
         etcd_servers: new_resource.etcd_servers,
@@ -57,7 +58,7 @@ action :create do
     template new_resource.master_file do
       source 'master.yaml.erb'
       variables(
-        erb_corsAllowedOrigins: new_resource.origins,
+        erb_corsAllowedOrigins: new_resource.origins + [node['cookbook-openshift3']['openshift_common_public_ip']] + node['cookbook-openshift3']['openshift_common_svc_names'],
         standalone_registry: new_resource.standalone_registry,
         erb_master_named_certificates: named_certificates,
         etcd_servers: new_resource.etcd_servers,